p2p: extract interface from v1 pipeline bus
