+++ logstash-core/src/main/java/org/logstash/plugins/pipeline/PipelineInput.java
@@ -40,6 +40,8 @@ public interface PipelineInput {
      */
     ReceiveResponse internalReceive(Stream<JrubyEventExtLibrary.RubyEvent> events);
 
+    String getId();
+
     /**
      * @return true if the input is running
      */
