+++ logstash-core/src/main/java/org/logstash/plugins/pipeline/PipelineOutput.java
@@ -24,4 +24,5 @@ package org.logstash.plugins.pipeline;
  * Represents the out endpoint of a pipeline to pipeline communication.
  * */
 public interface PipelineOutput {
+    String getId();
 }
